Fix various warnings. Improve OSWRCH call detection when saving traces.
Fix up b2 code as required. Hide one deprecation warning for an API
that was deprecated after b2's minimum macOS target version.

Silence warnings for some b2 dependencies. SDL deprecation warnings are
probably unavoidable. libuv type truncation warnings are probably
benign (it's a well used library). dylibbundler might be worth fixing,
but later (see #367).

+++ src/shared/c/system_osx.cpp
@@ -241,7 +241,7 @@ char **GetBacktraceSymbols(void *const *addresses, int num_addresses) {
      * very clear what a good alternative is, though? The following
      * should at least not go wrong too badly... */
 
-    size_t dyld_index = 0;
+    uint32_t dyld_index = 0;
     for (;;) {
         const char *image_name = _dyld_get_image_name(dyld_index);
         const struct mach_header *image_header = _dyld_get_image_header(dyld_index);
